DataRecovery fixed incorrect connection and match organization by name in tickets organizations.
git-svn-id: https://teamsupport.svn.beanstalkapp.com/main/trunk@3824 02885c34-e171-44ab-8435-42fb7a3c3aaf

+++ DataRecovery/Form1.cs
@@ -365,6 +365,8 @@ AND t.DateCreated < '2015-09-17 05:56:00'";
       command.Parameters.AddWithValue("@OrganizationID", orgID);
       badTickets.Fill(command, "");
 
+      Organizations existingCompanies = new Organizations(GetGoodLoginUser());
+      existingCompanies.LoadByOrganizationID(orgID);
 
       foreach (Ticket badTicket in badTickets)
       {
@@ -457,14 +459,15 @@ AND t.DateCreated < '2015-09-17 05:56:00'";
         }
 
 
-        Organizations orgs = new Organizations(GetGoodLoginUser());
+        Organizations orgs = new Organizations(GetCorrupteLoginUser());
         orgs.LoadBTicketID(badTicket.TicketID);
 
         foreach (Organization org in orgs)
         {
-          if (org.ParentID == orgID)
+          Organization goodCompany = existingCompanies.FindByName(org.Name);
+          if (org.ParentID == orgID && goodCompany != null)
           {
-            goodTicket.Collection.AddOrganization(org.OrganizationID, goodTicket.TicketID);
+            goodTicket.Collection.AddOrganization(goodCompany.OrganizationID, goodTicket.TicketID);
             EmailPosts.DeleteImportEmails(GetGoodLoginUser());
 
           }
